+++ antsor/lab_02/rectangle.cpp
@@ -9,7 +9,7 @@ Rectangle::Rectangle()
 	angle = 0;
 }
 
-Rectangle::Rectangle(Point p1, Point p2, double ang = 0)
+Rectangle::Rectangle(Point p1, Point p2, int ang = 0)
 {
 	tlp.setX(p1.x());
 	tlp.setY(p1.y());
@@ -18,7 +18,7 @@ Rectangle::Rectangle(Point p1, Point p2, double ang = 0)
 	angle = ang;
 }
 
-Rectangle::Rectangle(double x1, double y1, double x2, double y2, double ang = 0)
+Rectangle::Rectangle(int x1, int y1, int x2, int y2, int ang = 0)
 {
 	tlp.setX(x1);
 	tlp.setY(y1);
@@ -39,7 +39,7 @@ void Rectangle::setTopLeft(Point p)
 	tlp.setY(p.y());
 }
 
-void Rectangle::setTopLeft(double x, double y)
+void Rectangle::setTopLeft(int x, int y)
 {
 	tlp.setX(x);
 	tlp.setY(y);
@@ -57,18 +57,18 @@ void Rectangle::setBottomRight(Point p)
 	brp.setY(p.y());
 }
 
-void Rectangle::setBottomRight(double x, double y)
+void Rectangle::setBottomRight(int x, int y)
 {
 	brp.setX(x);
 	brp.setY(y);
 }
 
-double Rectangle::angle()
+int Rectangle::angle()
 {
 	return angle;
 }
 
-void Rectangle::setAngle(double ang)
+void Rectangle::setAngle(int ang)
 {
 	angle = ang;
 }
